+++ src/Laraplans/Contracts/SubscriptionResolverInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Gerardojbaez\Laraplans\Contracts;
+
+use Illuminate\Database\Eloquent\Model;
+
+interface SubscriptionResolverInterface
+{
+    /**
+     * Resolve the subscribable subscription.
+     *
+     * @param  Model  $subscribable
+     * @param  string $name The subscription name if your site supports multiple subscriptions.
+     * @return Model
+     */
+    public function resolve(Model $subscribable, $name) : Model;
+}