+++ src/Laraplans/LaraplansServiceProvider.php
@@ -4,10 +4,12 @@ namespace Gerardojbaez\Laraplans;
 
 use Illuminate\Support\ServiceProvider;
 use Gerardojbaez\Laraplans\SubscriptionBuilder;
+use Gerardojbaez\Laraplans\SubscriptionResolver;
 use Gerardojbaez\Laraplans\Contracts\PlanInterface;
 use Gerardojbaez\Laraplans\Contracts\PlanFeatureInterface;
 use Gerardojbaez\Laraplans\Contracts\PlanSubscriptionInterface;
 use Gerardojbaez\Laraplans\Contracts\SubscriptionBuilderInterface;
+use Gerardojbaez\Laraplans\Contracts\SubscriptionResolverInterface;
 use Gerardojbaez\Laraplans\Contracts\PlanSubscriptionUsageInterface;
 
 class LaraplansServiceProvider extends ServiceProvider
@@ -48,6 +50,7 @@ class LaraplansServiceProvider extends ServiceProvider
         $this->app->bind(PlanSubscriptionInterface::class, config('laraplans.models.plan_subscription'));
         $this->app->bind(PlanSubscriptionUsageInterface::class, config('laraplans.models.plan_subscription_usage'));
         $this->app->bind(SubscriptionBuilderInterface::class, SubscriptionBuilder::class);
+        $this->app->bind(SubscriptionResolverInterface::class, SubscriptionResolver::class);
     }
 
     /**