+++ src/Laraplans/SubscriptionResolver.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Gerardojbaez\Laraplans;
+
+use Illuminate\Database\Eloquent\Model;
+use Gerardojbaez\Laraplans\Contracts\SubscriptionResolverInterface;
+
+class SubscriptionResolver implements SubscriptionResolverInterface
+{
+    /**
+     * @inherit
+     */
+    public function resolve(Model $subscribable, $name) : Model
+    {
+        $subscriptions = $subscribable->subscriptions->sortByDesc(function ($value) {
+            return $value->created_at->getTimestamp();
+        });
+
+        foreach ($subscriptions as $key => $subscription) {
+            if ($subscription->name === $name) {
+                return $subscription;
+            }
+        }
+    }
+}